After activating your theme, you'll notice that it didn't look that great. That's because it wasn't including the stylesheet of the parent theme. We're now adding a function to the parent to enqueue both stylesheets.
We could have copied the style.css file into the child theme, but then we will have lots of duplicate code and styles that we might not need. We could also import the child theme in the stylesheet, but the filesystem might not necessarily be setup for that.

We're using wp_enqueue_style() to load from both the parent theme and the child theme.